requirements.txt
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning (optional)
# tensorflow>=2.8.0
# torch>=1.11.0

# Data processing
openpyxl>=3.0.0
xlrd>=2.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
click>=8.0.0
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# MLOps (optional)
# mlflow>=1.20.0
# wandb>=0.12.0
